edWith_pr7 (Volley를 이용해 API 통신 성공 , API 요청 -> 응답,  응답받은 데이터로 프래그먼트 생성 성공,  이미지 AsyncTask 이용해서 이미지 불러오기 성공   * Glide 이용해서 이미지 불러오기 성공     -> Glide 이용이 딜레이가 없어서 좋은거 같음        AsyncTask는 주석처리함
  * Detail 쪽 응답 받고 뷰에 값들 설정함

  * Detail 리뷰 모음 가져왔음

  * 리뷰 READ / WRITE 모두 데이터 타입에 맞춰서 수정함

  * API 서버에 한풀평 작성, 추천하기 완료함

  * API 서버에 좋아요&싫어요 버튼 & 구현함

  * Room을 이용해 데이터베이스 구현중 (DAO, Entity,
    DataBase) 만들었고 이제 연동하는 과정 중

  * Room 말고 Realm 사용해서 데이터베이스 비었을 때만 네트워크 사용 데이터베이스에 파일이 존재하면 파일을 사용하도록 했다.

  * MovieDetail 모델을 RealmObject 추가시 에러 발생 -> Realm 구조 이해 필요

  * 에러 발생 원인은 Realm은 모델이 추가되거나 모델의 구조가 변경될때 마이그레이션이 필요하다 마이그레이션없이 진행하려면 오류가 뜬다 그래서 개발할 때에는 Realm 객체를 생성할때 config(설정) 을 마이그레이션이 필요할 시 알아서 삭제해 주는 설정을 해 둔다.

  * DB 추가 완료 (메인메뉴, 영화 상세화면, 리뷰)

+++ edWithPR/app/src/main/java/luv/zoey/edwith_pr/MovieDetail/ReviewWriteActivity.kt
@@ -24,7 +24,7 @@ class ReviewWriteActivity : AppCompatActivity() {
             var reviewDTO = MovieReviewDTO(
                 movieID!!,                               // ID
                 getString(R.string.userNickName),                                // NAME
-                movieName,                                  // MOVIE NAME
+                movieName.toInt(),                          // MOVIE NAME
                 null,                            // IMAGE URL
                 System.currentTimeMillis().toString(),      // TIME
                 null,                              // TIME STAMP
